+++ src/shared/config/texts.ts
@@ -143,7 +143,6 @@ export const TEXTS = {
         categoryLabel: POST_FORM_COMMON.categoryLabel,
         privateLabel: POST_FORM_COMMON.privateLabel,
         privateDescription: POST_FORM_COMMON.privateDescription,
-        updating: '수정하는 중...',
         update: '수정하기',
       },
     },
@@ -291,6 +290,7 @@ export const TEXTS = {
     success: {
       accountCreated: '가입을 완료했어요.',
       postCreated: '포스트를 생성했어요.',
+      postUpdated: '포스트를 수정했어요.',
       linkCopied: '링크를 복사했어요.',
       originalLinkCopied: '원본 링크를 복사했어요.',
       bookmarkSavedTo: (folderName: string) => `${folderName}에 저장했어요.`,